+++ bin/actions/0220022-17.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+#: BR + W - Hit CTRL+W (close window)
+
+xdotool sleep 0.2
+xdotool key Control_L+w
+