+++ CHANGELOG.md
@@ -11,7 +11,7 @@ V0.3.1 - Not released yet
 * Add `help` and `h` command: List and explain available commands
 * Improve copy action (◣ + C): Auto-close popup, notify copied string
 * Improve BR + Up action: add jump to top in gnome-terminal
-~~~~
+
 V0.3.0 - 2020/12/20
 -------------------
 * Add options for using specific non-default keyboard device